--- TheatricalPlayersRefactoringKata/StatementPrinter.cs
@@ -1,40 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Globalization;
-using System.Linq;
-using System.Text;
-using System.Xml.Linq;
-
-namespace TheatricalPlayersRefactoringKata
-{
-    public class StatementPrinter
-    {
-        private readonly CultureInfo _cultureInfo = CultureInfo.GetCultureInfo("en-US");
-
-        private readonly CalculatorPrice calculatorPrice = new CalculatorPrice();
-
-        public string Print(Invoice invoice, Dictionary<string, Play> plays)
-        {
-            var result = new StringBuilder();
-            result.AppendFormat("Statement for {0}\n", invoice.Customer);
-
-            var totalAmount = 0m;
-            var volumeCredits = 0;
-
-            foreach (var perf in invoice.Performances)
-            {
-                var play = plays[perf.PlayId];
-                var thisAmount = calculatorPrice.CalculateAmount(play, perf);
-                volumeCredits += calculatorPrice.CalculateVolumeCredits(play, perf);
-
-                result.AppendFormat(_cultureInfo, "  {0}: {1:C2} ({2} seats)\n", play.Name, thisAmount, perf.Audience);
-                totalAmount += thisAmount;
-            }
-
-            result.AppendFormat(_cultureInfo, "Amount owed is {0:C2}\n", totalAmount);
-            result.AppendFormat("You earned {0} credits\n", volumeCredits);
-
-            return result.ToString();
-        }
-    }
-}
